modify all structure generate import write key

+++ mainwindow.h
@@ -46,14 +46,14 @@ private:
         RsaSSA::ModeName,
     };
     QList<QString>* m_aesKeys = new QList<QString>{
-        QString::number(KeyLength::LENGTH_DEFAULT),
-        QString::number(KeyLength::LENGTH_32),
-        QString::number(KeyLength::LENGTH_64),
-        QString::number(KeyLength::LENGTH_128),
-        QString::number(KeyLength::LENGTH_256),
-        QString::number(KeyLength::LENGTH_512),
-        QString::number(KeyLength::LENGTH_1024),
-        QString::number(KeyLength::LENGTH_2048),
+        QString::number(static_cast<int>(KeyLength::LENGTH_DEFAULT)),
+        QString::number(static_cast<int>(KeyLength::LENGTH_32)),
+        QString::number(static_cast<int>(KeyLength::LENGTH_64)),
+        QString::number(static_cast<int>(KeyLength::LENGTH_128)),
+        QString::number(static_cast<int>(KeyLength::LENGTH_256)),
+        QString::number(static_cast<int>(KeyLength::LENGTH_512)),
+        QString::number(static_cast<int>(KeyLength::LENGTH_1024)),
+        QString::number(static_cast<int>(KeyLength::LENGTH_2048)),
     };
     QList<QString>* m_encodings = new QList<QString>{
         "Base64",
